Extract Agent Token
This creates the new agent/credential/resource files and extract the
Agent Token. Because the Agent is now capturing the actor and no longer
dealing with Token concerns, the Conversion has been inlined again in
clean form.
