+++ lib/checkpoint/credential.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'checkpoint/credential/token'
+
+module Checkpoint
+  class Credential
+  end
+end
+
