+++ lib/checkpoint/resource.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'checkpoint/resource/token'
+
+module Checkpoint
+  class Resource
+  end
+end
+
